+++ blueprints/ember-cli-bootstrap-bookingsync-sass/index.js
@@ -1,0 +1,12 @@
+module.exports = {
+  normalizeEntityName: function() {},
+
+  beforeInstall: function() {
+    return this.addAddonsToProject({
+      packages: [
+        { name: 'ember-cli-sass', target: 'latest' },
+        { name: 'ember-cli-bootstrap-sassy', target: 'latest' }
+      ]
+    });
+  }
+};